requirements.txt
google-generativeai==0.8.5
gTTS==2.5.4
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.46.0

